+++ NEWS.md
@@ -4,7 +4,8 @@ Pentobi Release Notes
 Version 28.1-dev (v28-fixes)
 ----------------------------
 
-*
+* Autosave state whenever window becomes inactive, which retains
+  the state when logging off on Gnome wthout closing the window
 
 Version 28.0 (28 Sep 2025)
 --------------------------